chore(ui): rename src/constants/path_routes.ts -> src/constants/routePaths.ts
