+++ ui/src/constants/routePaths.ts
@@ -10,10 +10,19 @@ export const ROUTE_PATHS = {
     INDEX: '/profile',
     SETTINGS: '/profile/settings',
   },
+
+  ADMIN: {
+    INDEX: '/admin',
+    DASHBOARD: '/admin/dashboard',
+    USERS: '/admin/users',
+    SESSIONS: '/admin/sessions',
+    SETTINGS: '/admin/settings',
+  },
 } as const;
 
 // Type for all route paths
 export type AppRoutePath =
   | (typeof ROUTE_PATHS)[keyof typeof ROUTE_PATHS]
   | (typeof ROUTE_PATHS.AUTH)[keyof typeof ROUTE_PATHS.AUTH]
-  | (typeof ROUTE_PATHS.PROFILE)[keyof typeof ROUTE_PATHS.PROFILE];
+  | (typeof ROUTE_PATHS.PROFILE)[keyof typeof ROUTE_PATHS.PROFILE]
+  | (typeof ROUTE_PATHS.ADMIN)[keyof typeof ROUTE_PATHS.ADMIN];
