+++ ui/src/hooks/useAuth.ts
@@ -6,7 +6,7 @@ import {
   type UserRegister,
   UsersService,
 } from '@/client';
-import { ROUTE_PATHS } from '@/constants/path_routes';
+import { ROUTE_PATHS } from '@/constants/routePaths';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { useNavigate } from '@tanstack/react-router';
 import { useState } from 'react';
